requirements.txt
streamlit
torch
torchvision
pillow
opencv-python
numpy
requests
ultralytics
omegaconf
opencv-python-headless
moviepy==1.0.3
imageio-ffmpeg